+++ eoreader/reader.py
@@ -64,11 +64,16 @@ class Constellation(ListEnum):
     S2 = "Sentinel-2"
     """Sentinel-2"""
 
-    S2_CLOUD = "Sentinel-2 stored on cloud"
+    S2_E84 = "Sentinel-2 stored on AWS and processed by Element84"
     """
-    Sentinel-2 stored on cloud
+    Sentinel-2 stored on AWS and processed by Element84:
+    - Element84: arn:aws:s3:::sentinel-cogs - https://registry.opendata.aws/sentinel-2-l2a-cogs
+    """
 
-    For now, obly the one created by Element84 are supported: https://stacindex.org/catalogs/earth-search#/43bjKKcJQfxYaT1ir3Ep6uENfjEoQrjkzhd2?t=3
+    S2_SIN = "Sentinel-2 stored on AWS and processed by Sinergise"
+    """
+    Sentinel-2 stored on AWS and processed by Sinergise:
+    arn:aws:s3:::sentinel-s2-l1c and arn:aws:s3:::sentinel-s2-l2a - https://registry.opendata.aws/sentinel-2/
     """
 
     S2_THEIA = "Sentinel-2 Theia"
@@ -204,8 +209,9 @@ class Constellation(ListEnum):
 CONSTELLATION_REGEX = {
     Constellation.S1: r"S1[AB]_(IW|EW|SM|WV)_(RAW|SLC|GRD|OCN)[FHM_]_[0-2]S[SD][HV]_\d{8}T\d{6}_\d{8}T\d{6}_\d{6}_.{11}",
     Constellation.S2: r"S2[AB]_MSIL(1C|2A)_\d{8}T\d{6}_N\d{4}_R\d{3}_T\d{2}\w{3}_\d{8}T\d{6}",
-    # Element84 : S2A_31UDQ_20230714_0_L2A
-    Constellation.S2_CLOUD: r"S2[AB]_\d{2}\w{3}_\d{8}_\d_L(1C|2A)",
+    # Element84 : S2A_31UDQ_20230714_0_L2A, Sinergise: 0 or 1...
+    Constellation.S2_E84: r"S2[AB]_\d{2}\w{3}_\d{8}_\d_L(1C|2A)",
+    Constellation.S2_SIN: r"\d",
     Constellation.S2_THEIA: r"SENTINEL2[AB]_\d{8}-\d{6}-\d{3}_L(2A|1C)_T\d{2}\w{3}_[CDH](_V\d-\d|)",
     Constellation.S3_OLCI: r"S3[AB]_OL_[012]_\w{6}_\d{8}T\d{6}_\d{8}T\d{6}_\d{8}T\d{6}_\w{17}_\w{3}_[OFDR]_(NR|ST|NT)_\d{3}",
     Constellation.S3_SLSTR: r"S3[AB]_SL_[012]_\w{6}_\d{8}T\d{6}_\d{8}T\d{6}_\d{8}T\d{6}_\w{17}_\w{3}_[OFDR]_(NR|ST|NT)_\d{3}",
@@ -269,7 +275,14 @@ MTD_REGEX = {
         "regex": r".*s1[ab]-(iw|ew|sm|wv)\d*-(raw|slc|grd|ocn)-[hv]{2}-\d{8}t\d{6}-\d{8}t\d{6}-\d{6}-\w{6}-\d{3}\.xml",
     },
     Constellation.S2: {"nested": 3, "regex": r"MTD_TL.xml"},
-    Constellation.S2_CLOUD: rf"{CONSTELLATION_REGEX[Constellation.S2_CLOUD]}\.json",
+    Constellation.S2_E84: rf"{CONSTELLATION_REGEX[Constellation.S2_E84]}\.json",
+    Constellation.S2_SIN: {
+        "nested": -1,  # File that can be found at any level (product/**/file)
+        "regex": [
+            r"metadata\.xml",  # Too generic name, check also a band
+            r"B12\.jp2",
+        ],
+    },
     Constellation.S2_THEIA: rf"{CONSTELLATION_REGEX[Constellation.S2_THEIA]}_MTD_ALL\.xml",
     Constellation.S3_OLCI: r"Oa\d{2}_radiance.nc",
     Constellation.S3_SLSTR: r"S\d_radiance_an.nc",
@@ -536,6 +549,9 @@ class Reader:
                     # SPOT-4/5 constellations
                     elif const in [Constellation.SPOT4, Constellation.SPOT5]:
                         sat_class = "spot45_product"
+                    elif const in [Constellation.S2_SIN]:
+                        sat_class = "s2_product"
+                        kwargs["is_sinergise"] = True
 
                     # Manage both optical and SAR
                     try: